refactor(substrate): fold ctx.rs into component.rs
`ComponentCtx` and `StateBundle` move from `crate::ctx` into
`crate::component`. The ctx is the runtime half of the same primitive
the `Component` struct lives on (it's what the `Store<ComponentCtx>`
inside `Component` carries), so co-locating them removes the standalone
two-struct-one-file module that had no other content.

External consumers update their import paths:

- `aether_substrate::ctx::ComponentCtx` → `aether_substrate::component::ComponentCtx`
- `crate::ctx::{ComponentCtx, StateBundle}` → `crate::component::{ComponentCtx, StateBundle}`

The `pub use ctx::ComponentCtx;` re-export in `aether-substrate`'s
crate root collapses into the existing `component::Component` re-export,
so consumers that use the crate-root path (e.g. `aether_substrate_bundle`)
stay unchanged.

The two `ctx::tests` (`unknown_recipient_bubbles_up_with_sender_mailbox`,
`unknown_recipient_without_outbound_warn_drops`) move into the existing
`component::tests` block.
