+++ crates/aether-substrate/src/component.rs
@@ -2,14 +2,230 @@
 // and the cached handles needed to deliver mail. Mail payloads are
 // written to the guest at a static `MAIL_OFFSET`; a guest-side
 // allocator is parked until an actual use case forces the question.
+//
+// Holds the `ComponentCtx` (per-component context stored as wasmtime
+// `Store` data) and `StateBundle` (ADR-0016 state-migration payload)
+// alongside the `Component` itself — the ctx is the runtime half of
+// the same primitive, so it lives here rather than in a separate
+// module.
+
+use std::cell::Cell;
+use std::sync::Arc;
 
 use wasmtime::{Engine, Linker, Memory, Module, Store, TypedFunc};
 
-use crate::ctx::{ComponentCtx, StateBundle};
-use crate::mail::{Mail, ReplyTarget};
-use crate::reply_table::{NO_REPLY_HANDLE, ReplyEntry};
+use crate::input::InputSubscribers;
+use crate::mail::{Mail, MailKind, MailboxId, ReplyTarget, ReplyTo};
+use crate::mailer::Mailer;
+use crate::native_transport::NativeTransport;
+use crate::outbound::HubOutbound;
+use crate::registry::{MailboxEntry, Registry};
+use crate::reply_table::{NO_REPLY_HANDLE, ReplyEntry, ReplyTable};
 
 const MAIL_OFFSET: u32 = 1024;
+
+/// ADR-0016 §3: opt-in state migration payload. The substrate owns the
+/// buffer from the moment `save_state` is called on the old instance
+/// until the bundle is handed to the new instance via `on_rehydrate`
+/// (or discarded if no successor consumes it). Both fields are opaque
+/// to the substrate — the component owns versioning and the byte layout.
+#[derive(Debug, Clone)]
+pub struct StateBundle {
+    pub version: u32,
+    pub bytes: Vec<u8>,
+}
+
+/// Per-component context stored as wasmtime `Store` data. Holds the
+/// sender's own `MailboxId`, a handle to the shared mail queue, and a
+/// handle to the registry so the `send_mail` host function can route
+/// without consulting the scheduler's internals.
+///
+/// Deliberately does NOT hold the scheduler's full shared state — doing
+/// so would create an Arc cycle through `Scheduler owns Actor, Actor
+/// owns Store<ComponentCtx>, ComponentCtx back to Scheduler`. By holding
+/// only `Arc<Registry>` and `Arc<Mailer>` the cycle is broken: neither
+/// of those owns any actor.
+pub struct ComponentCtx {
+    pub sender: MailboxId,
+    pub registry: Arc<Registry>,
+    pub queue: Arc<Mailer>,
+    /// ADR-0013: direct outbound handle so the `reply_mail` host fn
+    /// can address a specific Claude session without routing through
+    /// a well-known sink. Broadcast still goes through
+    /// `hub.claude.broadcast`; reply is the session-targeted twin.
+    /// `HubOutbound::disconnected` when no hub is attached — sends
+    /// silently drop, matching the broadcast semantics.
+    pub outbound: Arc<HubOutbound>,
+    /// ADR-0021 subscriber sets, shared with the platform-event
+    /// publisher in `main.rs`. `#[actor]`-decorated components
+    /// auto-subscribe every `K::IS_INPUT` handler kind by mailing
+    /// `aether.input.subscribe` from the init prologue the macro
+    /// prepends (ADR-0033 phase 3); `InputCapability` processes the
+    /// mail and mutates this set.
+    pub input_subscribers: InputSubscribers,
+    /// ADR-0013 + ADR-0017: handle→entry map populated by
+    /// `Component::deliver` whenever an inbound mail has a meaningful
+    /// reply target — a Claude session (`ReplyEntry::Session`) or
+    /// another component (`ReplyEntry::Component`). The guest
+    /// receives an opaque `u32` handle as the 4th param on its
+    /// `receive` shim and passes it back to `reply_mail`; the
+    /// substrate routes either over `HubOutbound` or back through
+    /// `Mailer` based on the variant.
+    pub reply_table: ReplyTable,
+    /// Set by the `save_state` host fn during `on_replace`. The
+    /// substrate extracts it after hooks return via
+    /// `Component::take_saved_state`. Never read by the guest —
+    /// rehydration reads from a scratch offset written by the
+    /// substrate, not from here.
+    pub saved_state: Option<StateBundle>,
+    /// Set by the `save_state` host fn when it rejects a call (1 MiB
+    /// cap exceeded, OOB pointer). ADR-0016 §4: a failing save aborts
+    /// the replace; the substrate checks this after `on_replace` and
+    /// surfaces the message back up the control plane.
+    pub save_state_error: Option<String>,
+    /// Set by the `init_failed_p32` host fn when the guest's `init`
+    /// returns `Err(BootError)`. Issue 525 Phase 4b / issue 531: the
+    /// substrate reads this after `init` returns non-zero and
+    /// surfaces the message in `LoadResult::Err { error }`. The guest
+    /// stages the bytes here and returns 1 from its `init` shim;
+    /// `Component::instantiate` turns the staged message into a
+    /// `wasmtime::Error` so the existing load-failure path in
+    /// `dispatch_load_component` reports it like any other
+    /// instantiation error. None on the success path.
+    pub init_failure: Option<String>,
+    /// Trampoline transport whose `wait_reply` the
+    /// [`crate::host_fns::wait_reply_p32`] host fn delegates to.
+    /// `Some` for ctx instances built by [`WasmTrampoline::init`]
+    /// (issue 634 Phase 4 PR 3 — `transport.wait_reply` is now the
+    /// single source of inbox / overflow / correlation-filter
+    /// truth); `None` for the test paths that build `ComponentCtx`
+    /// without a real trampoline (the host fn returns
+    /// [`crate::host_fns::WAIT_CANCELLED`] in that case, matching
+    /// the pre-Phase-4 "no inbox installed" disposition).
+    pub transport: Option<Arc<NativeTransport>>,
+    /// ADR-0042 correlation counter. Per-component (one
+    /// `ComponentCtx` per component instance). Holds the *next* id
+    /// to mint; `prev_correlation()` reads `counter - 1` to return
+    /// the last one minted. Starts at `1` so that `0` always means
+    /// "no correlation" (backward-compat sentinel for waits that
+    /// don't filter, and for `prev_correlation` before any send).
+    ///
+    /// `Cell` instead of `AtomicU64`: the component is single-
+    /// threaded (ADR-0038 actor-per-component), so the counter is
+    /// never touched from multiple threads.
+    correlation_counter: Cell<u64>,
+}
+
+impl ComponentCtx {
+    /// Build a fresh ctx with empty state-migration slots and an
+    /// empty sender table. Using this over the struct literal keeps
+    /// the private fields (reply_table, saved_state,
+    /// save_state_error) internal to the wiring — callers should
+    /// never set them directly.
+    pub fn new(
+        sender: MailboxId,
+        registry: Arc<Registry>,
+        queue: Arc<Mailer>,
+        outbound: Arc<HubOutbound>,
+        input_subscribers: InputSubscribers,
+    ) -> Self {
+        ComponentCtx {
+            sender,
+            registry,
+            queue,
+            outbound,
+            input_subscribers,
+            reply_table: ReplyTable::new(),
+            saved_state: None,
+            save_state_error: None,
+            init_failure: None,
+            transport: None,
+            correlation_counter: Cell::new(1),
+        }
+    }
+
+    /// Wire the trampoline's `NativeTransport` into the ctx so the
+    /// [`crate::host_fns::wait_reply_p32`] host fn can route through
+    /// it. Called by [`WasmTrampoline::init`] right after constructing
+    /// the ctx (and before `Component::instantiate`, since the host
+    /// fn closure captures the ctx via the wasmtime `Store` data
+    /// pointer at instantiation time, not at host-fn call time, so
+    /// installing later than that is fine).
+    pub fn install_transport(&mut self, transport: Arc<NativeTransport>) {
+        self.transport = Some(transport);
+    }
+
+    /// Mint the next correlation id and bump the counter. Private —
+    /// callers that want a correlation use `ComponentCtx::send`,
+    /// which mints internally and tags the outgoing mail.
+    fn mint_correlation(&self) -> u64 {
+        let id = self.correlation_counter.get();
+        self.correlation_counter.set(id + 1);
+        id
+    }
+
+    /// Return the correlation id used by the most recent
+    /// `ComponentCtx::send` call. The `prev_correlation_p32` host fn
+    /// surfaces this to the guest so sync wrappers know what to
+    /// filter on in `wait_reply_p32`. Returns `0` (the "no
+    /// correlation" sentinel) before any send has been made.
+    pub fn prev_correlation(&self) -> u64 {
+        // counter holds the *next* id to mint; subtract to get the
+        // last one. `.saturating_sub(1)` covers the pre-send case
+        // where counter is still `1` (initial) → returns `0`.
+        self.correlation_counter.get().saturating_sub(1)
+    }
+
+    /// Dispatch mail. If the recipient is a sink, the handler runs inline
+    /// on the caller's thread. Otherwise defer to the mailer, which
+    /// routes to the component's inbox, warn-drops dropped/unknown
+    /// mailboxes, or bubbles unknown ids up to the hub-substrate when
+    /// a `HubOutbound` is wired (ADR-0037).
+    pub fn send(&self, recipient: MailboxId, kind: MailKind, payload: Vec<u8>, count: u32) {
+        // ADR-0042: mint a fresh correlation_id for this send and
+        // stash it on `last_correlation` so `prev_correlation_p32`
+        // can return it to the guest. The minted id rides on the
+        // outgoing `ReplyTo.correlation_id`; the reply's echo
+        // (auto-routed by `Mailer::send_reply`) carries it back, and
+        // `wait_reply_p32` filters on it.
+        let correlation = self.mint_correlation();
+        let reply_to = ReplyTo::with_correlation(ReplyTarget::Component(self.sender), correlation);
+
+        if let Some(MailboxEntry::Closure(handler)) = self.registry.entry(recipient) {
+            let kind_name = self.registry.kind_name(kind).unwrap_or_default();
+            // Component-originated mail: the sender is this ctx's
+            // mailbox, so its registry name is the `origin` any
+            // sink cares about (ADR-0011), and the same mailbox id
+            // rides on `reply_to.target` so sink handlers that want
+            // to reply (ADR-0041's io sink is the motivating case)
+            // can route `*Result` back to this component via
+            // `Mailer::send_reply`.
+            let origin = self.registry.mailbox_name(self.sender);
+            handler(
+                kind,
+                &kind_name,
+                origin.as_deref(),
+                reply_to,
+                &payload,
+                count,
+            );
+            return;
+        }
+
+        // Component / dropped / unknown all funnel through `Mailer::push`:
+        // - Component (ADR-0017): mail enters the recipient's inbox with
+        //   `reply_to.target = Component(self.sender)` so
+        //   `Component::deliver` can allocate a Component-variant
+        //   `ReplyEntry`.
+        // - Dropped: warn-drops in `route_mail`.
+        // - Unknown (ADR-0037): bubbles up to the hub-substrate via
+        //   `MailToHubSubstrate`; the `source_mailbox_id` it carries is
+        //   recovered from `reply_to.target` when it's a Component
+        //   variant (warn-drops otherwise).
+        self.queue
+            .push(Mail::new(recipient, kind, payload, count).with_reply_to(reply_to));
+    }
+}
 
 /// Sentinel the ADR-0033 `#[actor]` dispatcher returns from
 /// `receive_p32` when mail arrives with a kind id the component has
@@ -649,5 +865,91 @@ mod tests {
         let mail = SubstrateMail::new(M(0), aether_data::KindId(0), vec![], 1);
         component.deliver(&mail).expect("deliver");
         assert!(rx.try_recv().is_err(), "no frame should have been sent");
+    }
+
+    /// ADR-0037 Phase 1 + Phase 2: when a component sends to a mailbox
+    /// id the local registry doesn't know, `ctx.send` defers to the
+    /// mailer, which emits an upstream `MailToHubSubstrate` frame
+    /// carrying the sender's mailbox id so the hub can build a
+    /// `ReplyTo::EngineMailbox` for the receiving component.
+    #[test]
+    fn unknown_recipient_bubbles_up_with_sender_mailbox() {
+        use crate::outbound::EgressEvent;
+
+        let (outbound, outbound_rx) = HubOutbound::attached_loopback();
+        let registry = Arc::new(Registry::new());
+        let sender = registry
+            .try_register_closure("client", Arc::new(|_, _, _, _, _, _| {}))
+            .expect("register client mailbox");
+
+        let mailer = Arc::new(Mailer::new());
+        mailer.wire(Arc::clone(&registry));
+        mailer.wire_outbound(Arc::clone(&outbound));
+
+        let ctx = ComponentCtx::new(
+            sender,
+            Arc::clone(&registry),
+            Arc::clone(&mailer),
+            outbound,
+            crate::input::new_subscribers(),
+        );
+
+        let unknown = MailboxId(0xDEADBEEF_u64);
+        let kind = aether_data::KindId(0xABCD_u64);
+        ctx.send(unknown, kind, vec![1, 2, 3], 1);
+
+        let event = outbound_rx.try_recv().expect("bubble-up event emitted");
+        match event {
+            EgressEvent::UnresolvedMail {
+                recipient_mailbox_id,
+                kind_id,
+                payload,
+                count,
+                source_mailbox_id,
+                ..
+            } => {
+                assert_eq!(recipient_mailbox_id, unknown);
+                assert_eq!(kind_id, kind);
+                assert_eq!(payload, vec![1, 2, 3]);
+                assert_eq!(count, 1);
+                assert_eq!(source_mailbox_id, Some(sender));
+            }
+            other => panic!("expected UnresolvedMail egress, got {other:?}"),
+        }
+    }
+
+    /// No hub wired (disconnected substrate, or the hub chassis
+    /// itself): unknown recipients still warn-drop — no crash, no
+    /// upstream frame.
+    #[test]
+    fn unknown_recipient_without_outbound_warn_drops() {
+        let (outbound, outbound_rx) = HubOutbound::attached_loopback();
+        let registry = Arc::new(Registry::new());
+        let sender = registry
+            .try_register_closure("client", Arc::new(|_, _, _, _, _, _| {}))
+            .expect("register client mailbox");
+
+        let mailer = Arc::new(Mailer::new());
+        mailer.wire(Arc::clone(&registry));
+        // Deliberately no `wire_outbound`.
+
+        let ctx = ComponentCtx::new(
+            sender,
+            Arc::clone(&registry),
+            Arc::clone(&mailer),
+            outbound,
+            crate::input::new_subscribers(),
+        );
+
+        ctx.send(
+            MailboxId(0xDEADBEEF_u64),
+            aether_data::KindId(0xABCD),
+            vec![],
+            0,
+        );
+        assert!(
+            outbound_rx.try_recv().is_err(),
+            "no bubble-up without a wired outbound"
+        );
     }
 }
